fix: get market description for chart
